fix/parsser: comment at the end of line
fix/instra: compare as double (assert)

+++ src/core/Operands/Int32.cpp
@@ -7,7 +7,7 @@ Int32::Int32(std::string const & value): Operand(value, eOperandType::Int32) {
     try {
         int intValue = std::stoi(value);
 
-        _value = static_cast<int16_t>(intValue);
+        _value = static_cast<int32_t>(intValue);
         this->_stringValue = std::to_string(_value);
     } catch (const std::invalid_argument& e) {
         throw std::invalid_argument("ERROR: Invalid argument > Operand(Int32) ? Cannot convert [" + value + "] to integer.");
